influencer/requirements.txt
uvicorn==0.29.0
fastapi==0.110.0
pydantic==2.6.4
python-validator==0.0.8
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-multipart==0.0.9
aioredis==2.0.1
aiomysql==0.2.0
asyncpg==0.29.0
sqlalchemy==2.0.29
passlib==1.7.4
python-jose==3.3.0
aiofiles==23.2.1
celery==5.3.6
openai==1.25.1
psycopg2-binary
transitions
pytest
pytest-fastapi
pandas==2.2.2
openpyxl==3.1.2
